Observer_zerosixty returns times in seconds.

+++ src/observers/observer_zerosixty.h
@@ -12,8 +12,7 @@ extern observer_zerosixty_t *observer_zerosixty_new(
     unsigned target
 );
 
-/* gets the last 0-60 time, or 0 if not observed yet.
- * zerosixty in mircoseconds */
-extern void observer_zerosixty_get_zerosixty( observer_base_t *obs, unsigned *zerosixty );
+/* gets the last 0-60 time in seconds */
+extern void observer_zerosixty_get_zerosixty( observer_base_t *obs, float *zerosixty );
 
 #endif /* defined _INCLUDED_OBSERVER_ZEROSIXTY_H */
